+++ app/Http/Controllers/AdminUsersController.php
@@ -22,7 +22,7 @@ class AdminUsersController extends Controller
 
     public function create()
     {
-        $roles = Role::lists('name', 'id')->all();
+        $roles = Role::pluck('name', 'id')->all();
         return view('admin.users.create', compact('roles'));
     }
 
@@ -48,7 +48,7 @@ class AdminUsersController extends Controller
     public function edit($id)
     {
         $user = User::findOrfail($id);
-        $roles = Role::lists('name','id')->all();
+        $roles = Role::pluck('name','id')->all();
         return view('admin.users.edit', compact('user', 'roles'));
     }
 
@@ -69,7 +69,9 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
-        unlink(public_path().$user->photo->file);
+        if(file_exists(public_path().$user->photo->file)){
+            unlink(public_path().$user->photo->file);
+        }
         $user->delete();
         Session::flash('deleted_user', "The user is deleted successfully.");
         return redirect('/admin/users');
